+++ packages/world_countries/example/lib/ui/settings/widgets/sliders/shadow_offset_vertical_slider.dart
@@ -1,0 +1,22 @@
+import "package:flutter/material.dart";
+
+import "../../../../helpers/extensions/shadows_flag_theme_controller_extension.dart";
+import "settings_tile_slider.dart";
+
+final class ShadowOffsetVerticalSlider extends SettingsTileSlider {
+  const ShadowOffsetVerticalSlider(super.flagTheme, {super.key})
+    : super(
+        name: "Vertical Offset",
+        min: -3,
+        max: 3,
+        divisions: 30,
+        icon: Icons.vertical_align_center_rounded,
+      );
+
+  @override
+  double get value => flagTheme.shadowOffsetVertical;
+
+  @override
+  void handleChange([double? newValue]) =>
+      flagTheme.shadowOffsetVertical = newValue;
+}