+++ packages/world_flags/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 2.8.0
+
+NEW FEATURES
+
+- Improve Norfolk Island (NF) flag.
+
+REFACTOR
+
+- Added assert statements to `copyWith` methods for validation.
+- Updated flag properties from `isSimplified: true` to `isSimplified: null` across multiple countries.
+
+DOCUMENTATION
+
+- Added website with flags demo (to `pubspec.yaml` and `README.md`).
+- Enhanced documentation for isSimplified property with detailed explanations.
+
 ## 2.7.0
 
 NEW FEATURES
@@ -282,7 +298,6 @@ New full flags:
 - North Macedonia (MKD)
 - Barbados (BRB)
 - Eritrea (ERI)
-- Papua New Guinea (PNG)
 - Slovakia (SVK)
 - Tajikistan (TJK)